+++ neat_cache/lib/src/providers/redis.dart
@@ -13,61 +13,80 @@
 // limitations under the License.
 
 import 'dart:async';
-import 'dart:io' show SocketException;
-import 'package:dartis/dartis.dart';
+import 'dart:io' show IOException, Socket, SocketOption;
+import 'dart:typed_data';
+import 'resp.dart';
 import 'package:logging/logging.dart';
 import '../../cache_provider.dart';
 
-final _logger = Logger('neat_cache');
+final _log = Logger('neat_cache');
 
 class _RedisContext {
-  final Connection connection;
-  final Client client;
-  final Commands<String, List<int>> commands;
-  _RedisContext({this.connection, this.client, this.commands});
+  final RespClient client;
+  _RedisContext({
+    required this.client,
+  });
 }
 
-typedef _Op<T> = Future<T> Function(Commands<String, List<int>>);
-
 class RedisCacheProvider extends CacheProvider<List<int>> {
-  final String _connectionString;
+  final Uri _connectionString;
   final Duration _connectTimeLimit;
   final Duration _commandTimeLimit;
   final Duration _reconnectDelay;
 
-  Future<_RedisContext> _context;
+  Future<_RedisContext>? _context;
   bool _isClosed = false;
 
   RedisCacheProvider(
-    this._connectionString, {
+    Uri connectionString, {
     Duration connectTimeLimit = const Duration(seconds: 30),
     Duration commandTimeLimit = const Duration(milliseconds: 200),
     Duration reconnectDelay = const Duration(seconds: 30),
-  })  : _connectTimeLimit = connectTimeLimit,
+  })  : _connectionString = connectionString,
+        _connectTimeLimit = connectTimeLimit,
         _commandTimeLimit = commandTimeLimit,
         _reconnectDelay = reconnectDelay {
-    assert(_connectionString.isNotEmpty, 'connectionString must be given');
-    assert(!_connectTimeLimit.isNegative, 'connectTimeLimit is negative');
-    assert(!_commandTimeLimit.isNegative, 'commandTimeLimit is negative');
-    assert(!_reconnectDelay.isNegative, 'reconnectDelay is negative');
+    if (!connectionString.isScheme('redis')) {
+      throw ArgumentError.value(
+          connectionString, 'connectionString', 'must have scheme redis://');
+    }
+    if (!connectionString.hasEmptyPath) {
+      throw ArgumentError.value(
+          connectionString, 'connectionString', 'cannot have a path');
+    }
+    if (connectTimeLimit.isNegative) {
+      throw ArgumentError.value(
+          connectTimeLimit, 'connectTimeLimit', 'must be positive');
+    }
+    if (commandTimeLimit.isNegative) {
+      throw ArgumentError.value(
+          commandTimeLimit, 'commandTimeLimit', 'must be positive');
+    }
+    if (reconnectDelay.isNegative) {
+      throw ArgumentError.value(
+          reconnectDelay, 'reconnectDelay', 'must be positive');
+    }
   }
 
   Future<_RedisContext> _createContext() async {
     try {
-      _logger.info('Connecting to redis');
-      final connection = await Connection.connect(_connectionString)
-          .timeout(_connectTimeLimit);
+      _log.info('Connecting to redis');
+      final socket = await Socket.connect(
+        _connectionString.host,
+        _connectionString.port,
+      ).timeout(_connectTimeLimit);
+      socket.setOption(SocketOption.tcpNoDelay, true);
+
       // Create context
-      final client = Client(connection);
       return _RedisContext(
-        connection: connection,
-        client: client,
-        commands: client.asCommands(),
+        client: RespClient(socket, socket),
       );
+    } on RedisConnectionException {
+      throw IntermittentCacheException('connection failed');
     } on TimeoutException {
       throw IntermittentCacheException('connect failed with timeout');
-    } on SocketException {
-      throw IntermittentCacheException('connect failed with socket exception');
+    } on IOException catch (e) {
+      throw IntermittentCacheException('connect failed with IOException: $e');
     } on Exception {
       throw IntermittentCacheException('connect failed with exception');
     }
@@ -75,58 +94,53 @@ class RedisCacheProvider extends CacheProvider<List<int>> {
 
   Future<_RedisContext> _getContext() {
     if (_context != null) {
-      return _context;
+      return _context!;
     }
     _context = _createContext();
     scheduleMicrotask(() async {
       _RedisContext ctx;
       try {
-        ctx = await _context;
+        ctx = await _context!;
       } on IntermittentCacheException {
         // If connecting fails, then we sleep and try again
         await Future.delayed(_reconnectDelay);
         _context = null; // reset _context, so next operation creates a new
         return;
       } catch (e) {
-        _logger.shout('unknown error/exception connecting to redis', e);
+        _log.shout('unknown error/exception connecting to redis', e);
         _context = null; // reset _context, so next operation creates a new
         rethrow; // propagate the error to crash to application.
       }
       // If connecting was successful, then we await the connection being
       // closed or error, and we reset _context.
-      // Any errors have already been returned if any commands were affect
-      // otherwise the error is just a broken TCP connection and we ignore it.
       try {
-        await ctx.connection.done;
+        await ctx.client.closed;
       } catch (e) {
         // ignore error
       }
       _context = null;
     });
-    return _context;
+    return _context!;
   }
 
-  Future<T> _withRedis<T>(_Op<T> fn) async {
+  Future<T> _withResp<T>(Future<T> Function(RespClient) fn) async {
     if (_isClosed) {
       throw StateError('CacheProvider.closed() has been called');
     }
     final ctx = await _getContext();
     try {
-      return await fn(ctx.commands).timeout(_commandTimeLimit);
+      return await fn(ctx.client).timeout(_commandTimeLimit);
+    } on RedisCommandException catch (e) {
+      throw AssertionError('error from redis command: $e');
     } on TimeoutException {
       // If we had a timeout, doing the command we forcibly disconnect
       // from the server, such that next retry will use a new connection.
-      await ctx.connection.disconnect().catchError((e) {/* ignore */});
-
-      throw IntermittentCacheException(
-        'operation timed-out, we have closed the connection',
-      );
-    } on RedisConnectionClosedException {
-      await ctx.connection.disconnect().catchError((e) {/* ignore */});
-      throw IntermittentCacheException('connection already closed');
-    } on SocketException {
-      await ctx.connection.disconnect().catchError((e) {/* ignore */});
-      throw IntermittentCacheException('socket broken');
+      await ctx.client.close(force: true);
+      throw IntermittentCacheException('redis command timeout');
+    } on RedisConnectionException catch (e) {
+      throw IntermittentCacheException('redis error: $e');
+    } on IOException catch (e) {
+      throw IntermittentCacheException('socket broken: $e');
     }
   }
 
@@ -134,20 +148,55 @@ class RedisCacheProvider extends CacheProvider<List<int>> {
   Future<void> close() async {
     _isClosed = true;
     if (_context != null) {
-      final ctx = await _context.catchError((e) {/* ignore */});
-      if (ctx != null) {
-        await ctx.connection.disconnect().catchError((e) {/* ignore */});
+      try {
+        final ctx = await _context!;
+        await ctx.client.close();
+      } catch (e) {
+        // ignore
       }
     }
   }
 
   @override
-  Future<List<int>> get(String key) => _withRedis((redis) => redis.get(key));
+  Future<List<int>?> get(String key) => _withResp((client) async {
+        final r = await client.command(['GET', key]);
+        if (r == null) {
+          return null;
+        }
+        if (r is Uint8List) {
+          return r;
+        }
+        assert(false, 'unexpected response from redis server');
+
+        // Force close the client
+        scheduleMicrotask(() => client.close(force: true));
+      });
 
   @override
-  Future<void> set(String key, List<int> value, [Duration ttl]) =>
-      _withRedis((redis) => redis.set(key, value, seconds: ttl?.inSeconds));
+  Future<void> set(String key, List<int> value, [Duration? ttl]) =>
+      _withResp((client) async {
+        final r = await client.command([
+          'SET',
+          key,
+          value,
+          if (ttl != null) ...<Object>['EX', ttl.inSeconds],
+        ]);
+        if (r != 'OK') {
+          assert(false, 'unexpected response from redis server');
+
+          // Force close the client
+          scheduleMicrotask(() => client.close(force: true));
+        }
+      });
 
   @override
-  Future<void> purge(String key) => _withRedis((redis) => redis.del(key: key));
+  Future<void> purge(String key) => _withResp((client) async {
+        final r = await client.command(['DEL', key]);
+        if (r is! int) {
+          assert(false, 'unexpected response from redis server');
+
+          // Force close the client
+          scheduleMicrotask(() => client.close(force: true));
+        }
+      });
 }