+++ neat_cache/CHANGELOG.md
@@ -1,3 +1,9 @@
+## v2.0.1
+ * Fixed issue when adding multiple commands without waiting for them to complete first.
+
+## v2.0.0
+ * Migrated to null-safety, dropping dependency on `package:dartis`.
+
 ## v1.0.2
  * Upgrade `package:dartis`.